+++ vulnerable_people_form/form_pages/check_your_answers.py
@@ -1,20 +1,23 @@
-from flask import session, current_app
+from flask import session, current_app, redirect
 
 from .blueprint import form
 from .shared.constants import PostcodeTier
 from .shared.render import render_template_with_title
 from .shared.routing import route_to_next_form_page, dynamic_back_url
 from .shared.session import persist_answers_from_session, form_answers, get_summary_rows_from_form_answers, \
-    get_postcode_tier
+    get_postcode_tier, is_returning_nhs_login_user_without_basic_care_needs_answer
 from ..integrations import govuk_notify_client, spl_check
 
 
 @form.route("/check-your-answers", methods=["GET"])
 def get_check_your_answers():
     session["check_answers_page_seen"] = True
-    exclude_answers = None
-    if current_app.is_tiering_logic_enabled and get_postcode_tier() == PostcodeTier.VERY_HIGH.value:
-        exclude_answers = ['basic_care_needs']
+
+    if current_app.is_tiering_logic_enabled:
+        if is_returning_nhs_login_user_without_basic_care_needs_answer():
+            return redirect(append_querystring_params("/basic-care-needs"))
+        if get_postcode_tier() == PostcodeTier.VERY_HIGH.value:
+            exclude_answers = ['basic_care_needs']
 
     return render_template_with_title(
         "check-your-answers.html",
