+++ vulnerable_people_form/form_pages/view_answers.py
@@ -6,10 +6,7 @@ from .shared.render import render_template_with_title
 from .shared.session import (
     get_summary_rows_from_form_answers,
     get_postcode_tier,
-    is_nhs_login_user,
-    get_answer_from_form,
-    accessing_saved_answers
-)
+    is_returning_nhs_login_user_without_basic_care_needs_answer)
 
 
 @form.route("/view-answers", methods=["GET"])
@@ -19,7 +16,7 @@ def get_view_answers():
 
     if current_app.is_tiering_logic_enabled:
         if is_returning_nhs_login_user_without_basic_care_needs_answer():
-            return redirect("/basic-care-needs")
+            return redirect("/basic-care-needs?ca=1")
         if get_postcode_tier() == PostcodeTier.VERY_HIGH.value:
             exclude_answers.append('basic_care_needs')
 
@@ -27,12 +24,3 @@ def get_view_answers():
         "view-answers.html",
         summary_rows=get_summary_rows_from_form_answers(exclude_answers)
     )
-
-
-def is_returning_nhs_login_user_without_basic_care_needs_answer():
-    # Scenario: Where the postcode tier has increased to VERY_HIGH_PLUS_SHIELDING
-    # and no answer is present for 'basic_care_needs'
-    return is_nhs_login_user() \
-           and accessing_saved_answers() \
-           and get_postcode_tier() == PostcodeTier.VERY_HIGH_PLUS_SHIELDING.value \
-           and get_answer_from_form(["basic_care_needs"]) is None
