+++ vulnerable_people_form/templates/not-eligible-postcode-returning-user.html
@@ -15,6 +15,6 @@
 
   {{ govukButton({
         "text": "I've changed my address",
-        "href": append_querystring_params("/postcode-lookup")
+        "href": append_querystring_params("/postcode-lookup?ca=1")
       }) }}
 {% endblock %}
